Fix unknown-view condition for noticeView.
Each time noticeView was re-instantiated, I added a keybinding for it.
However, deleting a View doesn't delete its bindings, so they started
stacking up. The handler, in turn, didn't remove itself, and didn't
allow for ErrUnknownView- so it would panic when it was run a second
time.

Fix both ways:
- Make the closeView handler idempotent, i.e. fail quietly if it's
  ErrUnknownView.
- Add the binding at startup, rather than each time the Notice is
  instantiated. Since it's bound to the View, it *shouldn't* be
  activated when the View is nonexistant.