Refactor ModelView to export functions, skip threads.
I'm not entirely convinced this is better. Using channels rather than
methods to communicate in this way seems more idiomatic of Go; however,
it doesn't play super well with how gocui is structured.

The nice property of channels, that this doesn't provide, is that they
can be blocking or non-blocking based on the *client*'s choice; the
client can select{} away, or put the write in a background thread, or
whatever. In this case, it's always nonblocking, but completion etc.
happens under the hood.

So... what to do, to reconcile these? Recalling that we can send
channels along channels, hm. Notice and Message could return a
chan<- string, but a new one for each invocation; and the receiving end
is put in the closure that is eventually passed to the main thread. So
the main thread is what reads from the channel, and the client can put
the call in a goroutine or not as they see fit.

That's a lot of setup / teardown, though. The idea of having an
intermediate thread (at the ModelView layer, I guess?) is somewhat
appealing; how does that work? Not quite as I'd had previously, if we
want to make it syncable on the client end.

What if there aren't interrupts for filling content? That is, content
always comes from Layout. select{chan foo, default} means that the GUI
thread never blocks on the background for getting its data; for the
inverse (GUI sending to background threads), the infinite-buffer thing
is WAI.

That's unpleasantly equivalent to polling, though; moreover, it wont'
work. The main loop select{}s on UI events and Execute events; we have
to trigger an Execute anyway to get an update.

So... better to have the logic in closures, or the logic in Layout?
Layout feels much less racy- keeps a linear flow. But it means for
things like DeleteView we have to code that intent in some other way.
Also, think of the case where new messages are coming in from the
background faster than the UI can display: if Layout is *listening* on
that channel, then the UI becomes unresponsive to keyboard input.

Given that last point: NO, the main thread shouldn't block on *reading
from* unbuffered channels. That doesn't mean we can't get feedback on
"this is done", though- again, as long as it's buffered by another
goroutine.

...

So: Model wants to display. It calls ModelView.Foo(thing to display) and
gets back a <-chan struct{}, which will close when the write is
complete. ModelView.Foo an event in the event queue, whose last step is
to close the channel. *close* doesn't require a sync with the receiving
thread- so it's fine. (*sending* on a channel from the main thread is
risky; what if the background isn't ready to sync? I guess if it's a
buffered channel it's ok...)

The above paragraph could be implemented using context.WithCancel;
"Done" indicates that the action has been executed on the GUI thread.
While technically the same as what Context does, I think it's better to
*not* implement using Context; Context's verbiage is around
cancellation, not completion.

One could pass in a Context to the ModelView as well- allowing for
cancellation if the action is not completed on time, or if something
else supersedes it. I'll punt that until it's needed, though.