Use Model that notifies every 10s.
This is working, not as expected, but as intended, I guess. This is
indeed picking up the 'notified' signal- I just forgot that it's "notify
on appearance," not "notify on close." Meh, guess that's more things to
pass back if we want them.

But I'm not going to spend too much more time on this, as Notify isn't
something I'm really planning on carrying forward; this is more a useful
exercise to think / learn about concurrency patterns.

One bleh caveat- it looks like this *isn't* overwriting the existing
message when a new Notify comes in, so need to look at that.