Appear to make tests pass.
It makes sense that the timeouts would cause correctness issues- that's
a message expected to be sent on one end, that doesn't show up on the
other, if timeouts happen.
What makes less sense to me is why they would be so costly, w/rt making
the ticker look like it's going much faster.