Fix race condition within mock.UpdateCounter
This also updates controller_test in a couple of ways.
It accounts for a write that I hadn't previously accounted for,
and adds a commented-out-while debugging new test.

channel_test still fails, which is accurate- controller is picking up
the zero EventID as "last" after the resize. I think that's correct
until there's an AddEvent update.