Fix Update discipline
Make sure that Quit and other methods on UI controllers are called
within the UI thread.

It's probably possible to do without this- I was getting races from
reading HasQuit in RunSync, where the write wasn't necessarily in a
synced block. It would also be possible to move the Update logic into
the Controller itself- but that would obfuscate the right number of
calls to wait for in tests.

Bleh.