Don't fail/block CI on golang-at-tip
The 'tip' build is taking approximately twice as long as the 1.8/1.9
builds; it's basically all "getting the tools" (>100s vs <1s).

We're OK if we're broken on a new build for a while, and I'm getting
e-mails for it anyway; but we don't need to wait the extra several
minutes for that feedback.